Add structure for getting reed switch values [broken?]
Committing since it's probably useful.  But note that after these
changes, the 7-seg display is acting very oddly - it only displays
values up to around 0800 properly, beyond that it's a mess.  The code is
sending the right data strings to it, so I'm not sure what's up.
Perhaps it needs a factory reset?