Munge some real data
Hacked about with the old python2 server script to output some of its
data into the .h format, and allocate correct reed switch indices to
each of the sites on the map.  Program now runs with some real-ish data
for one day of the year.

Still bugged in that the data parser isn't pulling out the same labelled
column for each site - once that's fixed, we would ideally find the most
interesting day to use for the data - or, allow date selection.